- changed folder names, so that they don't start with capitals - added sub-folders client and persistence to make model structure more understandable
