+++ src/main/java/de/berlin/lostberlin/model/order/client/OrderShortDao.java
@@ -1,4 +1,4 @@
-package de.berlin.lostberlin.model.Order;
+package de.berlin.lostberlin.model.order.client;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -18,4 +18,3 @@ public class OrderShortDao {
     private long participantsNr;
     private String description;
 }
-
