+++ src/main/java/de/berlin/lostberlin/model/order/client/OrderStatusDao.java
@@ -1,4 +1,4 @@
-package de.berlin.lostberlin.model.Order;
+package de.berlin.lostberlin.model.order.client;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
