+++ src/main/java/de/berlin/lostberlin/model/order/persistence/DatePrefixedSequenceIdGenerator.java
@@ -1,4 +1,4 @@
-package de.berlin.lostberlin.model.Order;
+package de.berlin.lostberlin.model.order.persistence;
 
 import org.hibernate.HibernateException;
 import org.hibernate.MappingException;
