+++ src/main/java/de/berlin/lostberlin/model/order/persistence/StatusTypes.java
@@ -1,4 +1,4 @@
-package de.berlin.lostberlin.model.Order;
+package de.berlin.lostberlin.model.order.persistence;
 
 public enum StatusTypes {
     PENDING,
